update lha bench and remove eko bench script
