+++ benchmarks/lha_paper_bench.py
@@ -5,6 +5,7 @@ Benchmark to :cite:`Giele:2002hx` (LO + NLO) and :cite:`Dittmar:2005ed` (NNLO)
 import numpy as np
 from banana import register
 
+from eko.interpolation import make_lambert_grid
 from ekomark.benchmark.runner import Runner
 
 register(__file__)
@@ -36,11 +37,11 @@ class LHABenchmark(Runner):
     Globally set the external program to LHA
     """
 
-    external = "LHA"
-
-    theory = {}
-
-    rotate_to_evolution_basis = True
+    def __init__(self):
+        super().__init__()
+        self.external = "LHA"
+        self.theory = {}
+        self.rotate_to_evolution_basis = True
 
     def plain_theory(self, pto):
         """
@@ -116,7 +117,7 @@ class LHABenchmark(Runner):
                 {
                     "Q2grid": [1e4],
                     "ev_op_iterations": 10,
-                    # "debug_skip_singlet": True
+                    "interpolation_xgrid": make_lambert_grid(60).tolist(),
                 }
             ],
             ["ToyLH"],
@@ -134,33 +135,37 @@ class LHABenchmark(Runner):
 class BenchmarkVFNS(LHABenchmark):
     """Variable Flavor Number Scheme"""
 
-    theory = base_theory.copy()
-    theory.update(
-        {
-            "FNS": "ZM-VFNS",  # ignored by eko, but needed by LHA_utils
-            "kcThr": 1.0,
-            "kbThr": 1.0,
-            "ktThr": 1.0,
-            "nf0": 3,
-            "nfref": 3,
-        }
-    )
+    def __init__(self):
+        super().__init__()
+        self.theory = base_theory.copy()
+        self.theory.update(
+            {
+                "FNS": "ZM-VFNS",  # ignored by eko, but needed by LHA_utils
+                "kcThr": 1.0,
+                "kbThr": 1.0,
+                "ktThr": 1.0,
+                "nf0": 3,
+                "nfref": 3,
+            }
+        )
 
 
 class BenchmarkFFNS(LHABenchmark):
     """Fixed Flavor Number Scheme"""
 
-    theory = base_theory.copy()
-    theory.update(
-        {
-            "FNS": "FFNS",  # ignored by eko, but needed by LHA_utils
-            "NfFF": 4,
-            "nfref": 4,
-            "kcThr": 0.0,
-            "kbThr": np.inf,
-            "ktThr": np.inf,
-        }
-    )
+    def __init__(self):
+        super().__init__()
+        self.theory = base_theory.copy()
+        self.theory.update(
+            {
+                "FNS": "FFNS",  # ignored by eko, but needed by LHA_utils
+                "NfFF": 4,
+                "nfref": 4,
+                "kcThr": 0.0,
+                "kbThr": np.inf,
+                "ktThr": np.inf,
+            }
+        )
 
     @staticmethod
     def skip_pdfs(theory):
@@ -173,10 +178,58 @@ class BenchmarkFFNS(LHABenchmark):
         return ffns_skip_pdfs
 
 
+class BenchmarkRunner(BenchmarkVFNS):
+    """
+    Generic benchmark runner using the LHA VFNS settings
+    """
+
+    def __init__(self, external):
+        super().__init__()
+        self.external = external
+        self.sandbox = True
+
+    def benchmark_sv(self, pto):
+        """
+        Scale variations
+
+        Parameters
+        ----------
+            pto : int
+                perturbation order
+
+        Returns
+        -------
+            list(dict)
+                theory updates
+        """
+        high, low = self.sv_theories(pto)
+
+        # here we need to adjust also the
+        # apfel initial nf, which can't
+        # be accessed in other ways
+        if self.external == "apfel":
+            for sv_theory in [low, high]:
+                sv_theory["kcThr"] = 1.0 + 1e-15
+                sv_theory["EScaleVar"] = 0
+        low["XIR"] = np.sqrt(2.0)
+        high["XIR"] = np.sqrt(0.5)
+
+        self.run_lha([low, high])
+
+
 if __name__ == "__main__":
 
-    obj = BenchmarkVFNS()
-    # obj = BenchmarkFFNS()
+    # # Benchmark to LHA
+    # obj = BenchmarkVFNS()
+    # # obj = BenchmarkFFNS()
 
     # obj.benchmark_plain(2)
-    obj.benchmark_sv(1)
+    # # obj.benchmark_sv(2)
+
+    # VFNS benchmarks with LHA settings
+    programs = ["LHA", "pegasus", "apfel"]
+    programs = ["apfel"]
+    for p in programs:
+        obj = BenchmarkRunner(p)
+        # obj.benchmark_plain(2)
+        obj.benchmark_sv(2)
